# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API documentation for the News Aggregator application.",
        "contact": {
            "email": "mustafa.softcode@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a list of articles",
                "description": "Fetch paginated articles with optional filters such as keyword, category and source",
                "operationId": "93ded9e31119e6bbb22455c72ef97f19",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search articles by keyword in title or content",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter articles by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Filter articles by author",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter articles by source",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of paginated articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a single article by ID",
                "description": "Fetch a specific article by its ID",
                "operationId": "a85697b2464e9aaa3f9980a83edbef88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Single article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Article": {
                "title": "Article",
                "description": "An article object",
                "properties": {
                    "id": {
                        "description": "ID of the article",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the article",
                        "type": "string"
                    },
                    "body": {
                        "description": "Content of the article",
                        "type": "string"
                    },
                    "thumb": {
                        "description": "Thumb image of the article",
                        "type": "string"
                    },
                    "web_url": {
                        "description": "Web URL of the article",
                        "type": "string"
                    },
                    "category_id": {
                        "description": "Category of the article",
                        "type": "integer"
                    },
                    "source_id": {
                        "description": "Source of the article",
                        "type": "integer"
                    },
                    "news_source": {
                        "description": "API Source of the article",
                        "type": "string"
                    },
                    "published_at": {
                        "description": "Publication date of the article",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        }
    ]
}